Use defaultTimeZone when parsing string to date, in addition to when unparsing date to string.
This change only affects strings that do not have a timezone specified (neither Z nor a numeric offset). Previously, the string would be assumed to be in [NSTimeZone defaultTimeZone] and the parsed date would be adjusted to UTC from that timezone. Now, the date is assumed to be in ISO8601Parser's defaultTimeZone if it is non-nil, otherwise falling back on [NSTimeZone defaultTimeZone].

The only case where behavior will change is when parsing a date with no timezone specified in the string AND you've changed the defaultTimeZone property on ISO8601DateFormatter.