stringFromDate: should assign timeZone to the NSDateFormatter instance it creates. Otherwise the date is not properly converted from UTC to the requested timezone.
If you don't use ISO8601DateFormatter.defaultTimeZone, this change won't affect your code. If you do set defaultTimeZone, the formatter will now convert a passed-in date from UTC (which all NSDates are, by definition) to defaultTimeZone. Previously, it would not convert the date but would include the timezone specifier in the output string.